Converted to Android Studio Project with Gradle support
 com.github.paradam.infinitepager.InfiniteFragmentPagerAdapter
 com.github.paradam.support.v4.infinitepager.InfiniteFragmentPagerAdapter
Better support for updating FragmentPagerAdapter upon the data changing that could cause the ViewPager to show a different order of Fragments or a different set of Fragments altogether.

 com.github.paradam.infinitepager.InfiniteViewPager
Support for switching to adjacent Fragment at the other end of the list through the `setRelativeCurrentItem` methods.

 android.support.v4.view.SlidablePagerTitle
 android.support.v4.view.SlidableTabLayout
 android.support.v4.view.SlidableTabStrip
Based off Sample code from developer.android.com; SlidingTabColors (https://developer.android.com/samples/SlidingTabsColors/index.html), an implementation that supports the InfiniteViewPager and InfinitePagerAdapter.

The selected indicator being able to be at both the start and end of the tab list when swiping from end to start or start to end in an InfiniteViewPager.

SlidingTabLayout can be hidden and shown, with sliding up/down animations.

Similar to android.support.v4.view.PagerTabStrip, a `SlidableTabLayout` can be used as a child to android.support.v4.view.ViewPager with it being linked to the ViewPager automatically without needing to link in Java. Can also be linked by assigning `viewPager` to the id of the ViewPager within the same layout parent in the XML layout file.